+++ mobile/src/screens/SplashScreen.jsx
@@ -1,0 +1,33 @@
+// import { StatusBar } from 'expo-status-bar';
+import { StyleSheet, Text, View, Image, Button, Pressable } from 'react-native';
+// import { NavigationContainer } from '@react-navigation/native';
+// import { createNativeStackNavigator } from '@react-navigation/native-stack';
+// import { useFonts, Roboto_400Regular} from '@expo-google-fonts/roboto';
+// import SplashScreen2 from "./src/screens/splashScreen"
+
+
+function SplashScreen() {
+  return (
+    <View style={[{backgroundColor: "#D4DDDD"}, styles.container]}>
+      <Image
+        style = {styles.ystemLogo} source={require('../../assets/YStemLogo1.png')}
+      />
+    </View>
+  );
+  }
+
+export default SplashScreen;
+
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+
+  ystemLogo: {
+    width:327,
+    height:142
+  }
+});
